[Chore] 코드 구조 조정, GridPrefab 연결 해제돼서 다시 연결하기전에 저장
